requirements.txt
apache-libcloud==3.8.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
hupper==1.12
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
PasteDeploy==3.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
pycparser==2.21
pyramid==2.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
translationstring==1.4
urllib3==2.0.4
venusian==3.0.0
WebOb==1.8.7
zope.deprecation==5.0
zope.interface==6.0
